refactor: use all-reference-store in dashboard components
