+++ apps/web/src/services/dashboards/shared/dashboard-widget-container/WidgetViewModeModal.vue
@@ -11,7 +11,8 @@ import { cloneDeep, isEqual } from 'lodash';
 
 import { store } from '@/store';
 
-import type { AllReferenceTypeInfo } from '@/store/modules/reference/type';
+import type { AllReferenceTypeInfo } from '@/store/reference/all-reference-store';
+import { useAllReferenceStore } from '@/store/reference/all-reference-store';
 
 import { useManagePermissionState } from '@/common/composables/page-manage-permission';
 
@@ -45,12 +46,12 @@ const dashboardDetailStore = useDashboardDetailInfoStore();
 const dashboardDetailState = dashboardDetailStore.$state;
 const widgetFormStore = useWidgetFormStore();
 const widgetFormState = widgetFormStore.$state;
-
+const allReferenceStore = useAllReferenceStore();
 const state = reactive({
     widgetRef: null as WidgetComponent|null,
     hasManagePermission: useManagePermissionState(),
     currencyRates: computed(() => store.state.settings.currencyRates),
-    allReferenceTypeInfo: computed<AllReferenceTypeInfo>(() => store.getters['reference/allReferenceTypeInfo']),
+    allReferenceTypeInfo: computed<AllReferenceTypeInfo>(() => allReferenceStore.getters.allReferenceTypeInfo),
     component: null as AsyncComponent|null,
     initiated: false,
     variablesSnapshot: {} as DashboardVariables,
